css: Disable undersired drop feedback on views
While a view is accepting a drop hovering it, there is a blue ring, as
provided by libadwaita's stylesheet with a wildcard selector.

But this ring e partially hidden by the view's children and also clipped
by the window rounded corner.

There is a similar problem in list view cells, which get drop feedback
blue ring on each cell, instead of the whole row as we would prefer.

In all these cases, disable the blue ring with a specific selector.