name-cell: Improve drop feedback on list view
Since the GtkDropTarget is set on each cell, the default active state
(blue rectangle) is set only around a portion of the row in list view.
We disabled the GTK_STATE_FLAG_DROP_ACTIVE for view-cell anyways,
but we should have a better indication of drop state in list view since
the icons are small.

Heavily based on a different approach by Corey Berla