list-base: Support non-local drops
We reject drops on GtkDropTarget::enter if no drop action is supported.

But this assumes the drop value is already available, which holds only
for local (in-process) dnd opertaions. When dragging from other apps,
we need to preload the value first, which may happen at any point in
the future after ::enter.

So, as per GTK documentation's advice, listen to changes to the
GtkDropTarget:value property and update the cached preferred action
when a new non-NULL value is loaded.