sidebar: Open folder on DnD hover
Now that the sidebar is back under control of Nautilus, we
can handle hovering similar to how it's handled in the view.
While dragging across the sidebar, set a timeout when hovering over
a new row that accepts drops and is not the placeholder row
for new bookmarks.  Use the existing preference
NAUTILUS_PREFERENCES_OPEN_FOLDER_ON_DND_HOVER
to determine if folders should be opened on hover.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2056