sidebar: Change location on hover even when not a valid drop target
The "change location on hover" feature should work on any location
you can open, but it currently only works for valid drop targets.

Move the timer to the beginning of the motion handler so the
hover timer begins for locations that aren't valid drop targets.
This may cause hover_timer to receive a NULL row or a row
that doesn't have a valid uri so check for both of those
conditions in hover_timer().

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/1787