list-base: Clear scroll_to idle handler before clearing list-base
When you the navigation buttons to go back and forth very quickly
you can reliably get Nautilus to crash.  This happens because there
is an idle timer set to scroll to the item to be selected.  If you
change the view before the scroll has happened, the handler will
attempt to scroll to a non-existent item.

Clear scroll_to_file_handle_id before clearing all view items
in real_clear().