window: Don't create slot on construction
We create a view-less slot when the new window is constructed.

This is undesirable if the new window was created by detaching a tab,
because we already have got a slot (the one being detached); as a
workaround, we currently use a g_object_set_data() hack to close the
undesired tab.

But even in the other cases, this is entirely unnecessary, because
`nautilus_window_open_location_full()` creates the slot on demand!

So, let's just not create a slot on construction and wait for either
the detached tab or the open location.

Contrary to what an old comment states, the list of slots, and the
active_slot weak pointer can be NULL, as we perform NULL-checks
everywhere for them already.

(Note, tab detaching is currently disabled, but this
is a preparation to reenable it after porting to AdwTabView.)