window: Add slot in a single path
Currently, we add a slot to the window in 2 situations:

  1. When a new slot is created (`nautilus_window_initialize_slot()`);
  2. When a slot moves to another window (`notebook_page_added_cb()`).

However, `GtkNotebook::page-added` triggers in both cases, so we use
a hack with g_object_set_data() to know whether it happened by DND.

This hack is wrong because it catches only the cases when a tab is
detached into a new window, not when it's detached and attached to
an already existing window.

Fortunately, `nautilus_window_initialize_slot()` already calls
`nautilus_notebook_add_tab()`, which is going to cause an emission of
`GtkNotebook::page-added`. Therefore, we can simply add the slot
unconditionally and exclusively in `notebook_page_added_cb()`.

(Note, tab detaching is currently disabled, but this is a preparation
to reenable it after porting to AdwTabView.)