window: Remove "dnd-window-slot" hack
When a tab is moved into another window, we must remove the slot from
the old window.

Currently, we do this conditionally in `GtkNotebook::page-removed`.
The condition we use is a hack: we check for a boolean set as data.

This hack is wrong because it catches only the cases when a tab is
detached into a new window, not when it's detached and attached to
an already existing window.

So, instead, we can use a simpler condition: check whether the slot
has been removed from the slots list already. This way we can drop
the hack.

(Note: tab detaching is currently disabled, but this is a preparation
to reenable it after porting to AdwTabView.)