window: Simplify slot cleanup
We currently rely on sorting the list of slots in order to close the
active slot last when disposing of a window.

But the bug this works around only happens because we are actually
closing the tabs too. That's not necessary: widgets are going to be
destroyed later anyway along with their parents.