window: Refine slot removal code
`clost_slot()` does not necessarily closes the tab. This happens only
if the 3rd parameter is TRUE. This is misleading.

Now that there is only 1 case where the 3rd parameter is TRUE, let's
extract the tab-closing part out and rename the function accordingly.

Also, swap its parameters in order to make it `g_list_foreach()`-able.