file-utilities: Add ability to check for installed schema
Now that we are bringing the gnome-user-share extension
into Nautilus itself, we need the ability to check
if gnome-user-share is installed.  GNOME Settings
does this by checking the schema.  Let's use the same
code that Settings uses.

This doesn't work within sandboxes, so if we are
sandboxed always return TRUE;