special-location-bar: Extend to support Public folder for sharing
gnome-user-share maintains an extension to create an info bar
in the ~/Public folder notifying the user of the ability to turn
on file sharing.  This is broken with 43.alpha because of gtk3-gtk4.
This is a core feature and it will be easier to implement and
maintain in Nautilus itself rather than as an extension.

Extend special-location-bar for the Public folder / sharing.
Use dbus-launcher to open Settings to the "sharing" panel on
button click.

Fixes: https://gitlab.gnome.org/GNOME/gnome-user-share/-/issues/26
Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2155

See: https://gitlab.gnome.org/GNOME/gnome-user-share/-/merge_requests/18