grid-view: Don't set `focusable`
The view can receive focus, which causes it to draw an unnecessary
ring when focused and requires an extra focus step to get into the
items.

The view itself should not be focusable. This was set as a translation
of `can-focus`, but the semantics are different, and you generally
don't want to manually set `focusable` as the widgets set proper
defaults.