list-base: Move bg click gestures to view
We've been setting background click gestures on the view_ui widget,
which is either GtkGridView or GtkColumnView.

But this means the empty state widgets don't get to handle these
clicks. And the situation would get worse with AdwStatusPage, which
fills the whole area.

So, follow the lead of the background drop target and set the gestures
on the NautilusFilesView container itself. This has the side effect
of clicks on the floating status bar working like clicks on the empty
space. But it's not a big problem.