templates-dialog: Changes for the dialog box method
The current implementation of the “New Documents” submenu is a GtkPopoverMenu. It uses the GMenuModel, which is added as submenus in the right-click popover menu.
This implementation has many long-standing issues, as mentioned here https://gitlab.gnome.org/GNOME/nautilus/-/issues/2205, so we plan to move away from it by using a dialog box for creating new templates.

The dialog box makes it possible to view the whole list of templates in a single view, gives users the ability to search for templates, and makes it easier for everyone to use the template creation feature!