general: Refactor view icon from a GIcon to an icon name
Most APIs these days receive icons in the form of `icon-name`,
and we need to manually set up widgets like `GtkImage` if
we must use a GIcon. In the case of the view icon, we use
it for a button, which takes `icon-name`. So, this commit
reworks the API to use an icon name instead of GIcon.

Use an empty string as a fallback value while the slot has
no content view. This avoids trying to set NULL as icon-name.