location-entry: Add helper for setting editable text
We are setting the editable text on the entry, which in turn
calls the insert / delete text functions on the GtkText which
is invoking the signals we attached.  We only want the insert-text
and delete-text signals to be called when the user is actually
inputting data.  Add helper nautilus_location_entry_set_text()
which temporarily blocks the signals while setting text.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2253