location-entry: Drop DnD code
There's a complex set of code to handle DnD on the location entry.
This may have been very useful in the past, but it now a very corner
case situation.  Furthermore, the Editable has a DropTarget on it
already within Gtk.  The comments themself, indicate that likely
when a user is dropping multiple URI's it's a mistake.