toolbar: Factor out history controls
In order to adapt to narrow window sizes, we need to be be able to
show the same controls in different places, but we don't want to
duplicate the code or the UI definitions.

Thus, we should encapsulate the code and logic we need into distinct
reusable components.

Start with the back and forward buttons, along with their context menu.
They will update their state properly through bindings without needing
to call any functions ourselves.