toolbar: Set tooltip for view menu separately from toggle view
The toggle view button and view menu button are part of a
AdwSplitButton.  They both show a tooltip for "Show grid" / "Show List"
since the tooltip is set on the AdwSplitButton.

Set the tooltip for the AdwSplitButton as "View options" (for the
view menu) and overrid that tooltip for the toggle button using
adw_split_button_get_child()

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2327