app-chooser: Implement new dialog
When choosing to open a file in another application, the user may want
to set it as the new default.

Currently we require going to the Properties, which is not intuitive,
and one may think it's going to affect only one file.

Introduce a GtkAppChooserDialog replacement which provides the means to
set as default and reset as part of "Open with Another Application...".

Compared to the Properties "Open with" tab, the "Add" action is not
present, because opening the file with an app adds it implicitly. The
"Forget" action is not included either because it lacks its "Add"
counterpart and because such fine grained control is not essential. We
have Reset anyway.