rename-file-popover-controller: Set parent only on creation
We are attempting to reset the parent to the view every time
the rename popover shows.  This triggers warnings because the
popover already has a parent.  There's no reason to do this because
we are setting the parent to the view over and over again.
Since the popover and view have the same lifecycle, set the parent
on creation.