list-base: Update files view selection text when item removed
The selection text at the bottom on the view is usually updated
via the signal from the view-model GtkSelectionModel::selection-changed

This works except when items are removed, the signal emits before
the item is fully removed.  Rather than listening for all item changes
(i.e. GListModel::items-changed), update the selection text when
items are removed

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2338