files-view.c: Fixed delete key hint
The keypad hint for deleting a file displays as "Delete (keypad)"
instead of "Delete". It appears that this is due to the order in which
the Delete key and the Keypad Delete key are bound to the delete action.

In particular, for each of the five delete actions, the GDK_KEY_Delete
(0xffff) had been bound before the GDK_KEY_KP_Delete (0xff9f). As a
result, when one of those delete actions was later enabled, the most
recently bound key was used to populate the shortcut hint.

In this commit, each GDK_KEY_Delete binding is moved below the
corresponding GDK_KEY_KP_Delete binding so that the non-keypad delete
is the most recent binding, and thus the shortcut hint correctly
displays "Delete" instead of "Delete (keypad)".

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2357