pathbar: Align button menu with current mouse position
Right clicking parent folders in the path bar pops up the popover
in the middle of the button regardless of where the mouse is.
The popover becomes the child of the button creating the potential for
becoming an orphan.  Use the mouse coordinates with
gtk_popover_set_pointing_to() to set popover location.