Files-preferences-view: Fix terms inconsistency between UI and sources
Currently, there is a discrepancy between what the preferences menu
displays to the user and the denomination within the translatable
strings and related UI files.

Indeed, when hovering over the view menu, the user is shown the message
"Show grid", while "Icon View" is used within the preferences menu UI
file and, therefore, the translatable files.

To keep things consistent, "Icon" needs to be replaced by "Grid"
in the preferences menu UI file to be consistent with the actual
messages displayed in the application.

The aformentioned change needs to be applied to translatable strings
files by the localization team and tools.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2246.