files-view: Stop setting actions on submenus
We show Templates and Scripts submenus conditionally, using the
hidden-when=action-disabled attribute, by assigning a dummy action
which we can set as active or inactive as fit.

However, this causes criticals when the popover menu is destroyed:

(org.gnome.Nautilus:21502): GLib-CRITICAL **: 16:15:32.870: g_hash_table_iter_next: assertion 'ri->version == ri->hash_table->version' failed

Indeed, submenus are not supposed to have actions. There is a
"submenu-action" but it's for a different purpose.

So, instead of controlling the visibility through a dummy action,
set or unset the "hidden-when" attribute to control its visibility,
the same way we already do for sort menu items.