toolbar: Move "Visible Columns" menu item to folder menu
It would be nice to use gtk_column_view_column_set_header_menu(),
but unfortunately our GtkGestureClick for background context menu
overrules it. Also, it would be bad for accessibility because
there is no way to get keyboard focus in the column headers.

The folder menu is a fitting new home because:
   1. The column changes are saved only for this folder;
   2. If the user right clicks the columns, expecting an option to
      change columns, the option will be right there.

This is part of an effort to drop the view menu.