sidebar: Build with cloudproviders support
The cloudproviders integration used to be part of GTK 3 version, but
it is not part of GTK 4 version. This is because the sidebar codes
are now	part of	Nautilus codebase and `HAVE_CLOUDPROVIDERS` is never
set. Let's allow to build with cloudproviders support again.

Related: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2392