pathbar: Drop menu onboarding behavior
Before GNOME 42, the folder menu has been linked to the folder name for
a few releases, so users might have got used to click the current
folder name in order to access its menu.

With GNOME 42, the menu has been moved into a discrete button at the
end of the pathbar, as the "secondary menu" design pattern.

Anticipating people would click the current folder name out of habit,
and be frustrated that it didn't do anything, we've made this action
actually trigget the new menu button, in order to let users know of
the new location of the menu.

However, upon design review, this behaviour was deemed undesired.

Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2351