properties-window: Fix disproportionate icons
Thumbnails in Properties were sometimes larger than the requested size.

This is because we didn't pass the FORCE_THUMBNAIL_SIZE flag.

There were probably good reasons for this flag in the past, but there
are none now. If we request a size, that should be the size we get.

So, drop this flag entirely and default to forcing the size everywhere.