app-chooser: Put the AdwActionRow in a GtkListBox
eac32fcad6dc829a6388a7cc78a189dc83844a85 added an action row which
inherits from a GtkListBoxRow without putting it in a GtkListBox
(or something that inherits from GtkListBox like AdwPerferencesGroup).
This caused the activatable-widget not to work.  Put the row in a
GtkListBox since we don't want any of the other styles from Adw