app-chooser: Remove "reset" ability
We are only changing the default after clicking "open" based on the
design.  Resetting the default used to make sense when it has a
dedicated button, but it no longer makes sense.  Remove the "reset"
ability by setting the GtkSwitch as insenstive for the current
default and only taking action if the switch is active and sensitive.