dnd: Make "change location on hover" a little bit less aggressive
Reset the hover timeout if the x,y passes the GTK drag threshold
before the HOVER_TIMEOUT.  This makes it less likely that a location
will occur if the drag is merely happening over an item.  In GTK4
the cells cover a bigger space so this is even more likely to occur.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/1526