list-base: Prevent accidental hover-switch on slightest movement
When switching to a new folder by hovering it during DND, there is a
chance that a subfolder is going to appear right under the pointer.

Since the pointer is likely still moving (because it's tied to human
hand movement), the timeout for entering the subfolder starts right
away, resulting in an accidental location switch.

To prevent this, in the ::enter signal store the start coordinates
without starting the timeout. The ::motion handler is then going to
start the timeout only after significant pointer movement.

This has the side effect of also making it harder to accidentally
enter a folder by hovering only the edge of it, assuming that when
a person intends to open the folder on hover they are going to target
the center of the cell.