files-view: Make spacebar actually preview
We set spacebar as an action keybinding for previewing a file in sushi.

But this keybinding gets overriden by GtkListItemWidget for its own
`|listitem.select` action.

Maybe we shouldn't use the spacebar as previewer shortcut in the first
place, but that's how it's been so far. Also, we haven't been using
it for selection in previous versions either.

So, use a custom shortcut controller as a workaround.