dnd: Allow GDK_ACTION_COPY within the same folder
We are setting the preferred action to 0 when the source and dest are
in the same folder.  This prevents the ability to hold CTRL and perform
a copy operation (which worked in 42).  It isn't clear why you can't
override a 0 preferred action with key modifiers.

In order to keep consistency, remove this check in
nautilus_dnd_get_preferred_action() and add it to
nautilus_files_view_drop_proxy_received_uris() (only if the action
is MOVE).

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2438