files-view: Grab focus when background / selection menu closed
After closes the background or selection menu (by clicking out of the
menu or escape) the focus goes to the back button in the toolbar,
rather than the last focused item.  It's not clear if this is a
gtk issue or because our complex configuration.  Connect the "closed"
signal to simply grab back the focus to the files view.  This isn't
perfect because it doesn't necessarily grab the last focussed child,
but it saves from adding too much complexity.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2433