files-view: Set popover menu's parent to scrolled window
After closing the background or selection menu (by clicking out of the
menu or escape) the focus goes to the back button in the toolbar,
rather than the last focused item.  There's something related to
our FileView that isn't grabbing the focus back.  Set the parent
to the scrolled window instead so that we get the focus back.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2433