batch-rename-dialog: Fix default sort order
The list should be sorted by name-ascending, that state is set
when the action is created.  The sort is set when the action state
is changed, and since the action state isn't changed after creation, the
sort never happens.  This leads to an inconsistent state where
the list is not sorted, but the button label indicates that the list
is sorted.

Change the numbering-order state to name-ascending after the actions
are created (even though that state is already set) in order to force
list sorting.
Also remove, the default label on the sort button in order
to prevent this issue in the future (the label gets set anyways when
the state is changed).