properties: Fix opening gnome-disks for root filesystem
With the introduction of nautilus-dbus-launcher, we made it so that
the current disk is selected in gnome-disks (when opening from the
properties window) by passing the disk identifier to --block-device.
This works fine for most mounted volumes, but breaks the ability to
open gnome-disks for the root filesystem because
g_file_find_enclosing_mount() doesn't return a mount for the root file
system (it only returns "user interesting locations").

So, use unix-specific API to handle that case (antoniof's ammendment).

As a 2nd fallback, just open Disks without selecting anything (but if
this ever happens, it's a bug: we probably shouldn't show the Disks
section in Properties in that directory).

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2447