list-base: Make view click pressed event controller better
Our view click handling has an unnecessary workaround to deal with
the propagation from item clicks.  We set the flag deny_background_click
when an item is clicked, and then clear it on release/stop.  This is a
little hard to follow and I think might be contributing to problems
such a quasi-activated state for some items.

Set the propagation phase to TARGET on the view.  This works out of the
box for the GridView.  ColumnView is implemented using a child
ListView, so we have to get that by looking at the child's silbing.