files-view: Drop broken Alt+Down shortcut
It pretends to be the counterpart to Alt+Up but, in reality, it's just
an alternative to Enter/Return. It builds the illusion of being the
reverse of Alt+Up by relying on the fact that the previously open child
folder is selected after navigating up.

As such, the illusion falls short in two situations:

  - It works only once: pressing Alt+Down twice won't undo 2 Alt+Ups.
  - If selection changes, then Alt+Down will open the newly selected
    item, not the previously visited folder.

So, unbind this shortcut from this action. The shortcut is going to
be reintroduced for an actual "win.down" action.