dnd: Temporary fix for getting the preferred action on an x11 drop
In x11 the GtkDropTarget is not correctly setting the preferred action
on the GdkDrop.  The GdkDrag is getting set correctly, so for the
meantime, until a solution is merged into Gtk, let's get the action
from the GdkDrag instead of the GdkDrop.

See: https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/4982
Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2457