properties: Don't propagate height of subpages
Generally, we want the properties window to be sized to fit the natural
height of the content of the main page.

However, the extensions subpage is propagating its natural height; in
some cases, this can lead to the properties page being the full screen
height in many cases.

The natural height propagation is set internally by AdwPreferencesPage,
so we have to drop its use here. It was probably a misuse anyway, given
this is a subpage. Internally it's just a scrolled window and a clamp,
so that's what it's gonig to be replaced with.