general: Set labelled-by for view cells
In gtk4, we need to set labelled-by to help a11y programs know what
it currently selected.

Unfortunately, the focus is on a GTK internal widget, so at the moment
this solution requires a hack: we mark the internal list item widget as
being labeled by our cell (which is, in turn, labeled by the filename label).

We are reporting this to GTK to work toward a better solution.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2460