list-base: Check if item_widget is ready before setting focus
We use a hack to set the focus on the item when we set the selection.
Sometimes, the item_widget isn't ready by the time we are setting
the focus, which causes a seg fault.  This became even worse, when
I applied the focus hack more consistently in 2ac420316bd9c8560fa2cdeba0e5f5b0092243b7

We should find a better solution, but in the meantime, let's stop
crashing.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2400