sidebar: Make cloudproviders popover menu nested
libcloudproviders provides a menu_model that we use to build a
popover menu.  It's possible (and true for nextcloud) that these
menus can have sub-menus.  If the sub-menus have much longer
(or shorter) item names, the entire menu will get stretched out.
Use GTK_POPOVER_MENU_NESTED when creating the popover menu, creating
a nested menu which does not exhibit this issue.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2405