image-properties: Actually show failure info
We ask to append a failure message with NULL value, and then don't
actually append it because of the value being NULL. This makes no
sense, so fix it.

Despite that, in order to improve visuals, move the failure message
to the value and use an existing generic error string as title.