window-slot: Fix leak
Currently, if the new_selection list is NULL, then the
check_select_old_location_containing_folder function sets it to a newly
allocated list.

Since we don't own the new_selection list, let's copy it before passing
it to the check_select_old_location_containing_folder function so that
it always returns a newly allocated list, which we make the
pending_selection point to in order to avoid a leak.