list-base: Restore ability to move selection from sushi
Sushi emits a SelectionEvent signal when the arrow keys are pressed.
We listen to this signal to change selection in the view accordingly.

Unfortunately, the new view widgets do not implement a "move-cursor"
keybinding signal, so this is not working.

Partially resolve this regression by implementing the selection move
ourselves. This has a change in behaviour, however: up and down keys
do not move up and down in a grid, but instead move to the previous
and next items, respectively.