files-view: Re-work and simplifiy clipboard retrieval
The clipboard handling in files-view (and list-base) is centered
around NAUTILUS_CLIPBOARD which handles our needs locally, but
does not allow for flexibility around non-local types
(namely G_TYPE_FILE and GDK_TYPE_PIXBUF).  It's going to be
very hard / messy to make this work with new types, so let's
re-work the code to be more straightforward.