vfs-file: Handle setting metadata with a NULL list
Previous to 013edfaa7e567560baf2718ec75980b0118c93da, we had a special case
for handling NULL values when calling nautilus_file_set_metadata_list,
but after that commit, the NULL was passed directly to
g_file_info_set_attribute_stringv() which is not allow.  This broke
the "Reset to Default" button in the column chooser.

g_file_info_remove_attribute() does not produce the expected behavior
so simply set the attribute string to an empty string.