search-directory-file: Check for NULL before calling callback
Prevent segfault when calling call_when_ready with a NULL callback

This is the case with sidebar row, dc2711bd74056ed1711c0bce8dfcd9d0056bc3ad
which attempts to get the mount information from a file.  vfs-file
chains up to nautilus-directory and handles the NULL callback fine,
but if somehow, the search-directory was added to the bookmarks
(which isn't possible in current versions of Nautilus, but maybe
was in past version), Nautilus crashes.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2513