files-view: Only update the background menu on clipboard owner changed
Every time the clipboard owner is changed, we are updating the entire
context menus.  This is file with small folders, but if you have
100,000 items selected and you simply switch windows, it will
recalculate the entire selection menu.  To make matters worse, if you
receive a popup that the operation is taking a long time to complete,
that is a clipboard owner change, and it will recalculate again.
The only operation that we need from on_clipboard_owner_changed,
is whether to paste or not, so just update the actions state.