properties-window: Capitalize multi-byte characters correctly
Pango library doesn't support specifying text-transform in utf8
codepoints. It instead uses positions in the raw string. This is why we
have problems with, for example, the Cyrillic label in the UI file
text_transform positions are specified as 0 for the start and 1 for
the end and the first Cyrillic letter consist of two bytes.

Remove attributes from the UI file and implement label text capitalization
using the function eel_str_capitalize() which does the same but in
a proper way for type_value_label and type_file_system_label.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2494