ui: Use AdwBreakpoint to manage adaptive UI
AdwBreakpoint is a new utility that is part of
AdwWindow and AdwApplicationWindow. It allows for arbitrary
changes at specific breakpoints, meaning that we can move,
show, and hide widgets normally without any undefined behavior
or flickering.

AdwBreakpoint allows us to remove the GtkRevealers that we
used previously, as there's only a simple crossfade transition
between states now.