column-chooser: Don't allow "name" column to be hidden
During the population of the TreeView for the column-chooser,
the "sensitive" property disallows toggling the toggle for
the "Name" column.  The ::row-activated signal on the TreeView,
however, allows the "Name" column to be toggled because the
::row-activated signal is on the TreeView itself (rather than
the cell).  Either we could 1) remove ::row-activated altogether
(double clicking the display name) or 2) check to make sure
we aren't toggling "Name".  I chose #2.