list-view: Keep track of the most recently sorted column
In GtkColumnView, we don't know what column is being sorted on
from clicking on the column headers.  This means that we can only
save the sort order when the sort is intiated from the toolbar
menu rather than the column headers.  As a workaround, keep
track of the most recently used attribute used in the sorter.
Then when ::changed is emmited on the sorter, update the metadata
for what was used while sorting.  This is a pretty bad hack, but the
current state is worse because
1) The sort order is never saved in column view
2) If the sort order was changed (i.e. in GridView), the user
can't figure out why the order keeps reverting back to another state.