list-view: Only insert columns when they should be visible
In a nautilus-list-view, we create factories for the various
cells within each GtkListItem.  This is designed with performance in
mind... we don't want to create hundreds of thousands of objects
in a folder where we may only look at a small number of files.

We are proactively creating every possibly LabelCell without
checking which one is visible by appending all of the columns in
setup_view_columns.  This means that even though
the GtkListView may only request a couple hundred ListItems,
we end up generating thousands of LabelCells.  There are currently
13 available columns of which 3 are enabled by default.

Insert the columns into the columnview only when they are
selected to be visible.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2498