progress-info-manager: Don't leak finished & cancelled infos
NautilusProgressInfoManager aquires a reference for every progress info
and keeps these references in a list. Upon destruction, it releases the
reference while unreffing each progress info.

However, when it removes finished and cancelled infos from the list, it
doesn't release the reference, effectively leaking it.

This bug is a old as the manager: a0f89473830575ae39a568608bc58f3f92c6f4a4